server/requirements.txt
flask
pypdfium2
google-generativeai
python-dotenv